+++ src/main/java/org/example/topprogramingstoriesdashboard/scraper/NewStoriesGateway.java
@@ -3,5 +3,5 @@ package org.example.topprogramingstoriesdashboard.scraper;
 import java.util.List;
 
 public interface NewStoriesGateway {
-    List<Integer> getNewStories();
+    List<Long> getNewStories();
 }
